Fixed filesize calc, object to int cast.

+++ scqc/nemo.py
@@ -42,7 +42,7 @@ class Download(object):
         rundict = dict(zip(ddf.run_id, ddf.file_url))
         runlist = list(ddf.run_id)
         self.log.debug(f'{len(runlist)} runs in project {proj_id}')
-        totaldiskspace = ddf.file_size.sum() * 1e-9
+        totaldiskspace = ddf.file_size.astype(int).sum() * 1e-9
         self.log.debug(f'Expected disk space for {proj_id} files is {totaldiskspace} GB')
         donelist = []        
         triedlist = []
